requirements.txt
streamlit
chromadb
openai
python-dotenv
tiktoken
markitdown
pymupdf
